Broken Home page into component and hooks
