+++ src/pages/Home/index.jsx
@@ -1,59 +1,18 @@
-import { baseURL } from "../../utilities/constants";
-import ProductCard from "../../components/ProductCard";
-import ProductCardLoader from "../../components/ProductLoader";
-import { HomeHeading, ProductGrid, PageWrapper, Main, FilterContainer, Filter, HiddenLabel } from "./styles";
+import { HomeHeading, PageWrapper, Main, FilterContainer, Filter, HiddenLabel } from "./styles";
 import { updateHead } from "../../utilities/updateHead";
 import Search from "../../components/Search";
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { fetchingProducts, fetchedProducts, fetchFailed } from "../../context/productsSlice";
+import { useSelector } from "react-redux";
+import Products from "../../hooks/Product";
+import ProductsGrid from "../../components/ProductsGrid";
 
-function Products() {
-  const dispatch = useDispatch();
-  const { data, isLoading, isError } = useSelector((state) => state.products);
-
-  //prevent constant reloading, also a means of storing for search component.
-  useEffect(() => {
-    if (data.length === 0) {
-      const getProducts = async () => {
-        try {
-          dispatch(fetchingProducts());
-          const res = await fetch(baseURL);
-          const data = await res.json();
-          dispatch(fetchedProducts(data));
-        } catch {
-          dispatch(fetchFailed());
-        }
-      };
-      getProducts();
-    }
-  }, [dispatch, data]);
-
-  if (isLoading) {
-    return (
-      <ProductGrid>
-        {[...Array(25)].map((e, i) => (
-          <ProductCardLoader key={i} />
-        ))}
-      </ProductGrid>
-    );
-  }
-
-  if (isError) {
-    return <div>Fetch failed, please refresh and try again. If this problem persists contact the administrator at admin@admin.com.</div>;
-  }
-
-  return (
-    <ProductGrid>
-      {data.map((product) => (
-        <ProductCard key={product.id} product={product} />
-      ))}
-    </ProductGrid>
-  );
-}
-
+/**
+ * Creates the homepage
+ * @returns
+ */
 export default function Home() {
   updateHead("TechBeauty | Home", "TechBeauty an e-commerce website, selling a range of items from beauty and fashion product to audio and computer equipment.");
+  const { data, isLoading, isError } = useSelector((state) => state.products);
+  Products();
 
   return (
     <Main>
@@ -77,7 +36,7 @@ export default function Home() {
             <option>Name, Z-A</option>
           </Filter>
         </FilterContainer>
-        <Products />
+        <ProductsGrid data={data} isLoading={isLoading} isError={isError} />
       </PageWrapper>
     </Main>
   );
